Refactors shader structure for abstraction
Moves OpenGL shader implementations to an Abstracts folder to improve code organization and abstraction.
This change allows for better separation of concerns and facilitates future expansion with different shader implementations.
Updates include paths and CMakeLists.txt to reflect the new directory structure.
