+++ Rendering/Shaders/Abstracts/include/ShaderSourceManagerOpenGL.hpp
@@ -3,7 +3,7 @@
 
 
 #include "pch.hpp"
-#include "ShaderTypes.hpp"
+#include "../../Support/include/ShaderTypes.hpp"
 
 
 namespace Andromeda::Rendering
